regression(Dockerfile): revert to multiple copy directives
In the previous commit, only one `COPY` directive was used to attempt to
copy over the src, migrations, and sqlx-data.json directories and files.
However, the behavior of `COPY` is such that if multiple directories are
passed as sources, it only copies the contents of the directories to the
destination folder.

Signed-off-by: Sophon96 <71684640+Sophon96@users.noreply.github.com>